+++ changelog.md
@@ -2,9 +2,13 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9482]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9490]: https://www.github.com/atecoder/dn.vi.ivi
 
 Current IVI Visa Compatibility: 8.0.1.9391
+
+## [8.1.9490] - 2025-12-24
+Access subsystem
+- Rename registration to enrollment in the API roster.
 
 ## [8.1.9482] - 2025-12-17
 Pith